SF Feature Request 3170998.  Patch by Darryl Pierce (mcpierceaim): Provides support for the uniqueidentifier column type in SQLServer.
The column type is handled via the UniqueIdentierType provided by the patch. This type marshalls a UUID value into a uniqueidentifier column and back.  Modified the MsSqlDataTypeFactory to correctly return this new type when it detects a UUID.  Also included are unit tests to verify some of the new code.